+++ libft.h
@@ -3,27 +3,25 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bahn <bahn@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: bahn <bahn@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/12/21 15:37:34 by bahn              #+#    #+#             */
-/*   Updated: 2020/12/23 16:49:13 by bahn             ###   ########.fr       */
+/*   Created: 2020/12/23 20:51:15 by bahn              #+#    #+#             */
+/*   Updated: 2020/12/23 21:10:51 by bahn             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef	LIBFT_H
+#ifndef LIBFT_H
 # define LIBFT_H
-
-#include <unistd.h>
-#include <stdlib.h>
-#include <stdio.h>
+# include <unistd.h>
+# include <stdlib.h>
 
 int		ft_strlen(char *str);
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size);
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size);
+size_t		ft_strlcpy(char *dest, char *src, size_t size);
+size_t		ft_strlcat(char *dest, char *src, size_t size);
 char		*ft_strchr(char *str, int c);
 char		*ft_strrchr(char *str, int c);
-char		*ft_strnstr(char *dest, char *src, unsigned int size);
-int		ft_strncmp(char *s1, char *s2, unsigned int n);
+char		*ft_strnstr(char *dest, char *src, size_t size);
+int		ft_strncmp(char *s1, char *s2, size_t n);
 int		ft_atoi(char *str);
 int		ft_isalpha(int c);
 int		ft_isdigit(int c);
@@ -32,7 +30,5 @@ int		ft_isascii(int c);
 int		ft_isprint(int c);
 int		ft_toupper(int c);
 int		ft_tolower(int c);
-
 char		*ft_strdup(char *src);
-
 #endif
